Introduce separate opinion tone styling and layout
